🚚(app-desk) rename files in features mails
rename MailContent.tsx into MailsContent.tsx
rename MailLayout.tsx into MailsLayout.tsx
