+++ src/frontend/apps/desk/src/features/mails/components/MailsLayout.tsx
@@ -4,7 +4,7 @@ import { Box } from '@/components';
 import { MainLayout } from '@/core';
 import { useCunninghamTheme } from '@/cunningham';
 
-export function MailLayout({ children }: PropsWithChildren) {
+export function MailsLayout({ children }: PropsWithChildren) {
   const { colorsTokens } = useCunninghamTheme();
 
   return (
