+++ src/frontend/apps/desk/src/features/mails/components/index.tsx
@@ -1,2 +1,2 @@
-export * from './MailContent';
-export * from './MailLayout';
+export * from './MailsContent';
+export * from './MailsLayout';
