+++ src/frontend/apps/desk/src/pages/mails/index.tsx
@@ -1,14 +1,14 @@
 import { ReactElement } from 'react';
 
-import { MailContent, MailLayout } from '@/features/mails';
+import { MailsContent, MailsLayout } from '@/features/mails';
 import { NextPageWithLayout } from '@/types/next';
 
 const Page: NextPageWithLayout = () => {
-  return <MailContent />;
+  return <MailsContent />;
 };
 
 Page.getLayout = function getLayout(page: ReactElement) {
-  return <MailLayout>{page}</MailLayout>;
+  return <MailsLayout>{page}</MailsLayout>;
 };
 
 export default Page;
